Porocilo, dodatek inventuri

+++ zaloga/views.py
@@ -320,6 +320,8 @@ def spremeni_vnos(request, tip_baze, pk):
             vnos.tip = request.POST.get('tip')
         if cena and cena != "" :
             vnos.cena = float(request.POST.get('cena'))
+        if tip_baze == "inventura":
+            vnos.inventurna_sprememba(stevilo)
         vnos.save()
         return redirect('baza', tip_baze = tip_baze, pk = pk)
 
@@ -329,10 +331,13 @@ def izbrisi_vnos(request, tip_baze, pk):
         vnos = Vnos.objects.get(pk = request.POST.get('pk'))
         baza = Baza.objects.get(pk = pk)
         if baza.status == "veljavno":
-            tip = vnos.tip
-            sestavina = Sestavina.objects.get(dimenzija = vnos.dimenzija)
-            vnos.sprememba.delete()
-            sestavina.nastavi_iz_sprememb(tip)
+            if tip_baze == "inventura":
+                vnos.inventurni_izbris()
+            else:
+                tip = vnos.tip
+                sestavina = Sestavina.objects.get(dimenzija = vnos.dimenzija)
+                vnos.sprememba.delete()
+                sestavina.nastavi_iz_sprememb(tip)
         elif baza.status == "aktivno":
             vnos.delete()
         return redirect('baza', tip_baze=tip_baze, pk = pk)
